test(turn-orchestrator): retire approval-overlap cases and route more tests through shared fakes
- _helpers/stateStoreIii.ts: richer state-store-backed fake used by
  on-record-written.e2e; captures every state::get / state::set /
  state::update + turn::* wake.
- on-record-written.e2e.test.ts: switch to the shared helper.
- function-awaiting-approval.test.ts: drop the allow→steering_check
  and undecided-parking cases (parallel-approval.e2e covers both
  end-to-end with its dedicated harness); keep the function_execute
  resume case which the e2e doesn't exercise. Route the local makeIii
  through the shared fakeIii responder.
- preflight.test.ts, steering.test.ts, run-transition.test.ts: drop
  ad-hoc trigger fakes in favour of fakeIii({ responder }).

937 tests passing (-5 vs main: 3 tautological state.test.ts cases
from the previous commit + the 2 retired here).
